+++ src/stravalib/tests/integration/test_client.py
@@ -927,7 +927,9 @@ def test_get_segment_efforts(client, mock_strava_api):
     """Test that endpoint returns data as expected."""
     mock_strava_api.get("/segment_efforts", n_results=4)
 
-    efforts = list(client.get_segment_efforts(segment_id=2345, athlete_id=12345))
+    efforts = list(
+        client.get_segment_efforts(segment_id=2345, athlete_id=12345)
+    )
 
     assert len(efforts) == 4
     assert efforts[0].name == "Alpe d'Huez"