+++ src/stravalib/tests/integration/test_client.py
@@ -3,6 +3,7 @@ import json
 import os
 import warnings
 from unittest import mock
+from unittest.mock import patch
 
 import pytest
 import responses
@@ -53,6 +54,32 @@ def default_request_params():
         "elapsed_time": "3600",
     }
     return default_request_params
+
+
+@pytest.fixture
+def raw_exchange_response():
+    """Expected response from the protocol exchange_code_for_token method."""
+    return (
+        {
+            "access_token": "123456",
+            "refresh_token": "789sdf987",
+            "expires_at": 1732417459,
+        },
+        {
+            "id": 10295934,
+            "username": "foo_bar",
+            "resource_state": 2,
+            "firstname": "Foo",
+            "lastname": "Bar",
+            "bio": "A bio",
+            "city": "City",
+            "state": "State",
+            "country": "Country",
+            "sex": "F",
+            "premium": True,
+            "summit": True,
+        },
+    )
 
 
 def test_get_athlete(mock_strava_api, client):
@@ -1084,3 +1111,60 @@ def test_get_activity_kudos(mock_strava_api, client):
     assert isinstance(kudoer_list[0], SummaryAthlete)
     assert len(kudoer_list) == 2
     assert kudoer_list[0].lastname == "Doe"
+
+
+@patch("stravalib.protocol.ApiV3.exchange_code_for_token")
+def test_exchange_code_for_token_athlete(
+    mock_protocol_exchange_code, client, raw_exchange_response
+):
+    """If a user requests athlete data, then it should be returned as a
+    SummaryAthlete object."""
+    mock_protocol_exchange_code.return_value = raw_exchange_response
+
+    access_info, summary_athlete = client.exchange_code_for_token(
+        client_id=123,
+        client_secret="secret",
+        code="temp_code",
+        return_athlete=True,
+    )
+
+    assert access_info["access_token"] == "123456"
+    assert summary_athlete.firstname == "Foo"
+
+
+@patch("stravalib.protocol.ApiV3.exchange_code_for_token")
+def test_exchange_code_for_token_missing_athlete(
+    mock_protocol_exchange_code, client, raw_exchange_response
+):
+    """If athlete=True but Strava modifies the api response and athlete return
+    is None client should only return AccessInfo object"""
+    mock_protocol_exchange_code.return_value = (raw_exchange_response[0], None)
+    access_info, athlete = client.exchange_code_for_token(
+        client_id=123,
+        client_secret="secret",
+        code="temp_code",
+        return_athlete=True,
+    )
+
+    assert athlete is None
+    assert access_info["access_token"] == "123456"
+
+
+@patch("stravalib.protocol.ApiV3.exchange_code_for_token")
+def test_exchange_code_for_token_no_athlete(
+    mock_protocol_exchange_code, client, raw_exchange_response
+):
+    """When athlete isn't True, only return AccessInfo Typed Dict."""
+    mock_protocol_exchange_code.return_value = raw_exchange_response
+
+    access_info = client.exchange_code_for_token(
+        client_id=123,
+        client_secret="secret",
+        code="temp_code",
+        return_athlete=False,
+    )
+
+    assert (
+        access_info["access_token"] == raw_exchange_response[0]["access_token"]
+    )
+    assert isinstance(access_info, dict)