+++ changelog.md
@@ -6,7 +6,7 @@
 
 - Add: Add more information about how our mock fixture works (@lwasser, #292)
 
-### Fixed 
+### Fixed
 
 - Fix: Update segment_efforts api and return warnings (@lwasser, #321)